+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+* jdbc: all extensions can now be used on either a DataSource or an existing Connection
+
 # 2.0.4
 * jdbc: let tryLock/unlock not eat exceptions; also log if unlock returns false
 * jdbc: insertBatch()/upsertBatch() now work properly with mixing of nullable and not nullable @JsonColumn